Improved activation workflow
In our first pass at user activation, the link in the activation email
goes to the Django backend and invokes a view that validates the
activation code, activates the user, logs them into the session, and
redirects to the front-end app.

Change to link to the front-end app, which will make an API call to
the backend with the activation code. The API call will validate the
activation code, activate the user, and return the auth token for the
API in the response.

(This is ready except for changing the activation link that is
sent to the user.)