Include temporary link to service in approval email
We don't have a public page yet where we display services,
but the email template we got from IRC includes a link to
the service. Temporarily include a link to the service
editing page, just so there's not a blank in the email.

The way I'm building the link won't work right locally but
should be okay on staging and production; since it's just
a temporary solution, I think that's good enough.