Shrink bundle.js using closure
Compile it down to bundle_min.js.  This reduced the size for me
from 1.9M to 300K.

The drawback is that our source maps are lost, making debugging
more difficult.  I'm not sure there's a good solution to that.
In theory we could apply closure directly to our source files instead
of to bundle.js, and have closure generate a map file, but our
code is written assuming browserify, not closure, and would require
changes for closure to be able to process them directly.