Refs #292 Add service type icons to model
Populate from icons in code tree frontend/images.

Switch to a test runner that will not result in this population
in a migration generating a new set of these files every time the
tests are run.

Add a 'icon_url' field when returning a service type in the API
so the front end can display the icon if it wants to.

Display the icons next to the service type in the list of services
page.