342 Fix layout at top of services list page
When rendered RTL, text started at middle of page. This was
apparently due to having a div class=half up there. I'm not
sure why that wasn't getting laid out at the leading side
of the page whichever direction it was being rendered, but
there was no need for a div class=half there anyway, so I
just got rid of it.

Fixes #342