359 Display multiline text correctly
Add a handlebars helper (template tag) that
adds an HTML `<br>` whereever the text had
a newline.

This turned out more complicated than I had hoped,
because the helper needs to be registered on the
handlebars compiler and we're precompiling so it
was tricky to get at the compiler being used. I
ended up adding a wrapper compiler and configuring
hbsfy to use that.

Fixes #359