468 Provide password change/reset
Add a button on the login page that goes to a password
reset page, where the user can enter an email and if that
email has an account, we'll send them a password reset link.
That link will bring up another page on the frontend where
they can enter a new password twice and we'll change that
account's password and log them in. The password form
only works with a valid key from the link that was emailed,
and the key can only be used once.

Fixes: #468