479 Update footer text
I couldn't find a quick way to get the first phrase bolded and still
be compatible with proper translation. We ought to embed something
in the English string so the translators can move the markup if the
translated language changes the word order.  But embedded HTML markup
gets escaped by i18n and appears literally on the page.

Let's just get the text updated for now, and maybe later can fix the
bolding.

Fixes #479