491 Don't blow up on bad location in search API
If the lat,long on the closest parameter isn't valid,
just ignore it instead of returning a 500.

This appears to be happening with a very old version of Firefox
(5.0) providing us with location data that's in a format we're
not expecting (or just bad), so it's beyond the user's control
and therefore I didn't want to return a 400 error for it.

Fixes #491