547 Reloading reports page doesn't work
This was because the router checks the config to see
if the user is a staff user before it'll allow going on
to the reports view, but whn we reload, we're triggering
the route before the config has finished initializing.
Fix: wait for config to be ready before checking for
staff and routing to the view.

Fixes #547