561 Add geo data to service areas
Created a new LebanonRegion model and imported the
geo data for the top two levels of Lebanon's administrative
boundaries, the governates and the cazas or districts.
Added a foreign key to ServiceArea that links to the
corresponding region record.

(I didn't want to just stick the geo data directly into
the service area record because we might want to re-run
the geo data import at some point and doing it this way
keeps that model separate, so we can just clear it and
re-do the import without breaking the service area data.)

Fixes: #561