563 Don't display reports button to non-staff
Fix bug due to localStorage converting everything to strings, so
that forever.isStaff = false became "false" which is true-ish.
Use JSON on everything going into localStorage instead.

Added a couple of front-end tests for staff & non-staff, verified
that it reproduced the bug without the fix, and passed after the
fix.

Fixes #563