579 Fix mobile keyboard popping up over map
Well, this fixes one case, anyway. We had code that always set the
focus on the search input field when the map page was loaded, which
resulted in the keyboard popping up. Apparently we don't need that
code anymore. Originally we were reloading the page on each map update
and losing the focus, but changes in map rendering appear to have made
this fix no longer necessary. The focus stays in the input field as
you type and the map is updated.

After this fix, our code doesn't change the focus at all, which is
good. But when navigating around the app under android, if focus is
in the input field, then android is going to keep popping up the
keyboard on the slightest excuse, no matter how many times you close
it again. The solution is to tap just outside the input field to get
the focus off of it. Apparently tapping and dragging on the map
itself leaves the focus alone, so you have to tap outside the map.

Fixes #579