Try to fix language toggle test failures
These are random, and apparently due to the local storage
sometimes retaining the selected language, so we don't get
the language selection popup when we first load the home
page in a test as we expect.

Try to fix by clicking the language change menu button if
we timeout looking for the popup, then trying again.