Responsive images
Closes #43

This adds an api field to Service which has URLs and widths for large,
medium, and small images. The maximum pixel width that we'll use is
about 640, but we include a large (1280) image for high density screens.

The position of the image has been moved to the bottom of the detail
screen.

The `srcset` attr specifies the image URLs and their width, and the
`sizes` attr specifies that we want to use 100% of the viewport. Those
specifications are then used by the browser to pick an appropriate
image. The `src` param is included, which allows for non-supported
browsers to fallback.