WIP: waiting time chart
This is a WIP. I'm keeping the table in place for the moment. You can
step through all the charts, but i need to make the rating charts do
averages instead of totals. Also I need to split the final chart into
2 (one for type and one for location).

But before I proceed further, just wanted to see if I was on the right path.