Fix 403s from API form submission
This is a theoretical fix because I cannot reproduce the issue locally.

The issue occurs in this piece of code:
https://github.com/tomchristie/django-rest-framework/blob/master/rest_framework/authentication.py#L124-L128

If an anonymous user hits that code, then it's OK because authenticate
will return None and not try to `enforce_csrf`.

If a logged-in user hits that code, then `enforce_csrf` will get hit and
will fail. (That failure puzzles me, but I don't quite understand how
CSRF validation is done in this project. I assume that it's different
because of the client-API nature of it)

I think reversing the order of the authentication classes will work
because DRF short circuits at the first class that succeeds. In the
anonymous user case, both will fail, but that's OK, and `enforce_csrf`
won't get hit. In the logged-in user case, TokenAuthentication will
succeed, so the SessionAuthentication will not get hit at all.

Addresses #56