Incorporating Google Analytics into Browserify setup
To better use the Google Analytics code from both sides of Service Info, it has
been moved into a Browserify module and `require()`ed from both sides.

The use of the path `../../../../frontend/` is pretty cumbersome, but until we
set up the Gulpfile with some path-related cleverness, this is what we're stuck
with (and I propose leaving that for a future branch).