Revising appearance of error messages
Right now, the #messages element is indeed being filled out with all relevant error messages. But because they're piled on top of one another, only the most recently added is visible.

This adjusts the style of #messages to allow all messages to be displayed, one on top of another (in the second dimension, not the third).

It also adds code to ensure that the page's margins are set to make the messages visible when they're displayed and to tuck content back in place after they're hidden.