Adding FAQ styles
According to Omar, the FAQ list should be styled in such a way as to group questions and answers into a box and to mark off the questions from the answers. This styling can be fairly simple.

This set of changes adds such styles. It also removes the links to full FAQ pages from the relevant template altogether -- it is unwanted and therefore not needed.