Adding clearfix styles
The issue with the column layout is that columns are laid out with a float-based layout system. This requires that the parent
element have a "clearfix" style—and indeed the element had a "clearfix" *class*, but because no stylesheet was included, a clearfix *style* remained to be implemented. This does that.