Menu stopgap measures
Hides overflow content from the nav-menu. No more ghostly nav items floating around beneath the nav.

Relocates the language picker to its own `ul` outside the nav `ul` (an easy way to ensure it has the same appearance and behavior). This `ul` is not allowed to resize or wrap. Thus the items in the real nav `ul` will be squeezed out of their row while the language picker remains stoically in place.