SC-36  add aldryn-video
This is not hosted on pypi, hence the installation from github.

I noticed some Python 2-isms in some of the code related to iframes,
but I haven't been able to reproduce the problem; it isn't normally
used.  If time allows I'll submit a pull request to the project and
not worry about integrating a fix into ServiceInfo at this time.

Notable advantages of aldryn-video over djangocms-video:

* working Vimeo support
* better widget for YouTube videos

I don't see any obvious need to mess with the included templates.