Keeping nav menus with active children open
The easiest way to ensure that collapsible navigation menus with active children remain open is to trigger their opening behavior after the page loads by triggering a `click` event on the collapsible menu headers.

Why not just set the `active` class on those headers prior to the initialization of the collapsible menus? As it turns out, this looks exactly the same as this does—that is, there is a brief moment where the menus are not visibly expanded—so the improvement doesn't seem significant. This would also require wedging a `script` block between the jQuery include and the include for Materialize's JS, which doesn't seem good either.