Better menu item padding
Right now, menu items are padded not using actual padding but just the `line-height` attribute. This works well … until they're more than a line long. And unfortunately this is often the case. Then they begin to bump nastily against the edges of their boxes.

This sets them up to use real, actual padding.

To test this, either create a page with a ridiculously long name and see how it looks in the menu (it should look like its menu item is very tall but nicely padded on the top and bottom) or otherwise manipulate the text content of an element in the dev tools to be unusually long.