Adding fullcalendar
Google's own iframe-based calendar embedding is unsatisfactory and non-responsive. It's better to use a third-party plugin that consumes the Google calendar API. That is what this branch adds.

The preliminary steps for testing out this addition are outlined in the [fullcalendar Google Calendar instructions page](http://fullcalendar.io/docs/google_calendar/). The steps can be summarized as follows:

1. Create a Google Calendar API Key.
2. Create a Google calendar and make it public.
3. Obtain the ID for the calendar.

The next step is specific to the implementation here. What you must do is add a snippet of HTML like the following to a text plugin (or any other plugin allowing raw HTML editing):

    <div class="fullcalendar" data-api-key="<your API key>" data-calendar-id="<your calendar ID>"></div>

This element, once added and saved, will *no longer be visible* in the text plugin's HTML source in future edits. To change or remove it, the entire text plugin must be removed.

Once the element has been added to the source, the page can be refreshed, and the JavaScript that initializes the calendar will find the element, pull out the necessary data from its `data-` attributes, and initialize the calendar. Any number of calendars can be set up this way; they all just have to have the `fullcalendar` class and the relevant data.

The calendar widget is set up to be sensitive to the page's language. The `request.LANGUAGE_CODE` value is used to pick out a separate localization JS file. No special RTL transform of the CSS is necessary; fullcalendar's style sheet is apparently smart enough to handle that on its own.