Removing flex layout from .row elements
In retrospect, it was completely unnecessary to include a flex layout override for Materialize's `row` elements.

The supposition I made was that rows' columns were not going to be reordered correctly on RTL mode. This was wrong. We're applying an RTL transform to the Materialize CSS, and that CSS specifies `float:left` for `col` elements, which becomes `float:right` after the transform is applied. This fixes row layout for RTL.