Adding .h-scroll class styles
For many elements, it's useful to be able to scroll past the horizontal extent of the viewable area. Tables are a good example.

Unfortunately, the easiest way to accomplish this is to ask the content creator to wrap these elements in a `<div class="h-scroll">` element, whose styles are set up in this patch.