Adding legacy deps to Gulp build process
In order to ensure that misplaced dependencies cause a fail at build time rather than runtime, it's a good idea to get the build process to copy those dependencies into some directory we control.

This way, if there's a problem with Node versioning that causes dependencies to end up somewhere unexpected, we'll know early on instead of waiting to see 404s.

In this case, the `fullcalendar` and `moment` dependencies are included in the gulpfile.

The `.gitignore` is amended to ignore the `service_info/static/lib` build target.